refactor: move calendar icon to own component
